Complete restructuring of the [trunk].
Since PyroShop is indeed a module, all of the files and folders were place in the canonical PyroCMS [dir] structure for shared_addons so as to allow for easier installtion of the PyroShop module and its dependencies.

By default, PyroShop already depends on the galleries module. In this commit, within the galleries's model [galleries_images_m] was updated to obtain thumbnail_id(s) for the PyroShop [view_item] view.

In short, we now have a new directory structure and updated gallery:
modules > shop > [all the files pertaining to the PyroCMS addon]
modules > galleries > [updates to the galleries module that are necessary for PyroShop to work properly]